Move VTabOpaqueCursor to vtab module
